LayoutTests/fast/js/kde/garbage-n-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Expected an identifier but found 'code' instead
KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Got expected error: 'SyntaxError'
PASS successfullyParsed is true

TEST COMPLETE
